Fix limit of result bug
As we were overriding the attribute results the limit was not actually
being applied. By now removing it from the list and enable cascade
delete on orphans sqlalchemy will remove it.

Note this is preferred over calling delete on the session as this way it
doesnt matter whether the object was persisted or not (not persisted
object should be removed with spurge rather than delete)