Reuse retrieve code when performing a toss
When a toss is done on a draw the resulting draw needs to be return.
Rather than relying on the local memory object, perform a full retrieve
from DB to ensure the user is getting the up to date version of the draw
and that it will see the same result as just retrieving it.