Split test per draw type into its own file
This commit splits the test for models into a folder of multiple files.
The common actions are moved into a conftest and basic model validations
are left into a "common" test file.
This allows us to create model specific test per file, which makes it
easier to validate and understand what is needed to create a new draw.