Use class based views to dispatch requests
Rather than having the mapping and the free functions, this commits
moves the implementation to flask class method based views.

This allows not only to better define the classes linked to a view or
draw type but also to customize specific verbs of a draw implementation.

Additionally, as the draw type is not a parameter anymore, a better
response is returned in the case where a draw type does not exist.